+++ client/src/pages/group_auth/utils/schema.ts
@@ -2,8 +2,27 @@ import * as yup from 'yup';
 import { AUTH_TYPES } from './types';
 
 //  remarks: form login
-
-export const login_form_schema: yup.ObjectSchema<AUTH_TYPES['login']> = yup.object({
+const login_form_schema: yup.ObjectSchema<AUTH_TYPES['login']> = yup.object({
   username: yup.string().required('Username is required'),
   _password: yup.string().required('Password is required'),
 });
+
+//  remarks: forget password (sending out reset url)
+const reset_opt_out_schema: yup.ObjectSchema<AUTH_TYPES['reset_pw_opt_out']> = yup.object({
+  input: yup.string().required('Username is required'),
+});
+
+//  remarks: reset password (via reset url)
+const reset_opt_in_schema: yup.ObjectSchema<AUTH_TYPES['reset_pw_opt_in']> = yup.object({
+  _password: yup.string().required('Password is required'),
+  _password_confirm: yup.string().required('Password confirmation is required'),
+});
+
+
+
+//  remarks: collection export
+export const AUTH_SCHEMA = {
+  login: login_form_schema,
+  reset_pw_opt_out: reset_opt_out_schema,
+  reset_pw_opt_in: reset_opt_in_schema
+}