+++ client/src/pages/group_auth/utils/types.ts
@@ -1,7 +1,9 @@
 //  remarks centralisation
 
 export type AUTH_TYPES = {
-  login: TLogin
+  login: TLogin,
+  reset_pw_opt_out: TResetPasswordOptOut,
+  reset_pw_opt_in: TResetPasswordOptIn
 }
 
 //  remarks: types
@@ -9,4 +11,13 @@ export type AUTH_TYPES = {
 type TLogin = {
     username: string,
     _password: string
+  }
+
+type TResetPasswordOptOut = {
+    input: string
+  }
+
+type TResetPasswordOptIn = {
+    _password: string,
+    _password_confirm: string
   }